+++ src/frontend/src/pages/WorkPackageDetailPage/WorkPackageViewContainer/WorkPackageDetails.tsx
@@ -3,10 +3,10 @@
  * See the LICENSE file in the repository root folder for details.
  */
 
-import { ChangeRequest, WorkPackage, wbsPipe } from 'shared';
+import { WorkPackage, wbsPipe } from 'shared';
 import { percentPipe, fullNamePipe, datePipe, weeksPipe } from '../../../utils/pipes';
 import WbsStatus from '../../../components/WbsStatus';
-import { Box, Divider, Grid, Link, Stack, Typography, useTheme } from '@mui/material';
+import { Box, Divider, Grid, Link, Stack, Typography } from '@mui/material';
 import DetailDisplay from '../../../components/DetailDisplay';
 import WorkPackageStageChip from '../../../components/WorkPackageStageChip';
 import { timelinePipe } from '../../../utils/pipes';
@@ -16,83 +16,13 @@ import StackedLineChartIcon from '@mui/icons-material/StackedLineChart';
 import DoneAllIcon from '@mui/icons-material/DoneAll';
 import { Link as RouterLink } from 'react-router-dom';
 import { routes } from '../../../utils/routes';
-import ChangeRequestDetailCard from '../../../components/ChangeRequestDetailCard';
-import { useAllChangeRequests } from '../../../hooks/change-requests.hooks';
-import LoadingIndicator from '../../../components/LoadingIndicator';
-import ErrorPage from '../../ErrorPage';
-import { projectWbsPipe } from '../../../utils/pipes';
+
 interface WorkPackageDetailsProps {
   workPackage: WorkPackage;
   dependencies: WorkPackage[];
 }
 
 const WorkPackageDetails: React.FC<WorkPackageDetailsProps> = ({ workPackage, dependencies }) => {
-  const theme = useTheme();
-  const { data: changeRequests, isError: crIsError, isLoading: crIsLoading, error: crError } = useAllChangeRequests();
-  const currentDate = new Date();
-
-  if (crIsLoading || !changeRequests) return <LoadingIndicator />;
-  if (crIsError) return <ErrorPage message={crError?.message} />;
-
-  const crUnreviewed = changeRequests
-    ? changeRequests
-        .filter((cr: ChangeRequest) => !cr.dateReviewed && projectWbsPipe(cr.wbsNum) === projectWbsPipe(workPackage.wbsNum))
-        .sort((a, b) => b.dateSubmitted.getTime() - a.dateSubmitted.getTime())
-    : [];
-
-  const crApproved = changeRequests
-    ? changeRequests
-        .filter(
-          (cr: ChangeRequest) =>
-            cr.dateReviewed &&
-            cr.accepted &&
-            currentDate.getTime() - cr.dateReviewed.getTime() <= 1000 * 60 * 60 * 24 * 5 &&
-            projectWbsPipe(cr.wbsNum) === projectWbsPipe(workPackage.wbsNum)
-        )
-        .sort((a, b) => (a.dateReviewed && b.dateReviewed ? b.dateReviewed.getTime() - a.dateReviewed.getTime() : 0))
-    : [];
-  const displayCRCards = (crList: ChangeRequest[]) => (
-    <Box
-      sx={{
-        display: 'flex',
-        flexDirection: 'row',
-        flexWrap: 'wrap',
-        overflow: 'auto',
-        justifyContent: 'flex-start',
-        '&::-webkit-scrollbar': {
-          height: '20px'
-        },
-        '&::-webkit-scrollbar-track': {
-          backgroundColor: 'transparent'
-        },
-        '&::-webkit-scrollbar-thumb': {
-          backgroundColor: theme.palette.divider,
-          borderRadius: '20px',
-          border: '6px solid transparent',
-          backgroundClip: 'content-box'
-        }
-      }}
-    >
-      {crList.map((cr: ChangeRequest) => (
-        <ChangeRequestDetailCard changeRequest={cr}></ChangeRequestDetailCard>
-      ))}
-    </Box>
-  );
-
-  const renderChangeRequests = (title: string, crList: ChangeRequest[], emptyMessage: string) => {
-    return (
-      <>
-        <Typography variant="h5" gutterBottom>
-          {title}
-        </Typography>
-        {crList.length > 0 ? (
-          <Grid container>{displayCRCards(crList)}</Grid>
-        ) : (
-          <Typography gutterBottom>{emptyMessage}</Typography>
-        )}
-      </>
-    );
-  };
   return (
     <>
       <Box sx={{ display: 'flex', flexDirection: 'row', alignItems: 'center', gap: '8px', mb: 2, mt: 1 }}>
@@ -143,16 +73,6 @@ const WorkPackageDetails: React.FC<WorkPackageDetailsProps> = ({ workPackage, de
           <ScheduleIcon sx={{ mr: 2 }} />
           <DetailDisplay label="Timeline Status" content={timelinePipe(workPackage.timelineStatus)} paddingRight={1} />
         </Grid>
-      </Grid>
-
-      <Grid
-        sx={{
-          mb: 1,
-          mt: 3
-        }}
-      >
-        {renderChangeRequests('Un-reviewed Change Requests', crUnreviewed, 'No un-reviewed change requests')}
-        {renderChangeRequests('Recently Approved Change Requests', crApproved, 'No recently approved change requests')}
       </Grid>
 
       <Typography
